requirements.txt
acme==0.26.1
alabaster==0.7.10
anaconda-client==1.6.14
anaconda-navigator==1.8.7
anaconda-project==0.8.2
APScheduler==3.5.1
asn1crypto==0.24.0
astroid==1.6.3
astropy==3.0.2
attrs==18.1.0
autobahn==18.8.1
Automat==0.7.0
azure-common==1.1.14
azure-nspkg==2.0.0
azure-storage==0.36.0
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
bokeh==0.12.16
boto==2.48.0
boto3==1.7.59
botocore==1.10.59
Bottleneck==1.2.1
cachetools==2.1.0
cairocffi==0.8.0
CairoSVG==2.1.3
certbot==0.26.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.4
conda==4.5.4
conda-build==3.10.5
conda-verify==2.0.0
ConfigArgParse==0.13.0
configobj==5.0.6
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.3.1
cssselect2==0.2.1
cx-Oracle==6.2.1
cycler==0.10.0
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.5
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
distributed==1.21.8
Django==2.0.5
django-easy-pdf==0.1.1
django-secure==1.0.1
dnspython==1.15.0
docutils==0.14
DoubleMetaphone==0.1
email-validator==1.0.3
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.4
future==0.16.0
gevent==1.3.0
glob2==0.6
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
graphviz==0.8.3
greenlet==0.4.13
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
httplib2==0.11.3
hyperlink==18.0.0
idna==2.6
ijson==2.3
imageio==2.3.0
imagesize==1.0.0
incremental==17.5.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
josepy==1.1.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.23.1
lml==0.0.1
locket==0.2.0
lxml==4.2.1
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
mock==2.0.0
more-itertools==4.1.0
mpmath==1.0.0
msgpack==0.5.6
msgpack-python==0.5.6
multipledispatch==0.5.0
nameparser==0.5.6
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
noise==1.2.2
nose==1.3.7
notebook==5.5.0
numba==0.38.0
numexpr==2.6.5
numpy==1.15.2
numpydoc==0.8.0
oauth2client==4.1.2
oauthlib==2.0.7
odo==0.5.1
olefile==0.45.1
openpyxl==2.5.3
packaging==17.1
pandas==0.22.0
pandocfilters==1.4.2
parsedatetime==2.4
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
pbr==4.1.1
pdfkit==0.6.1
pdfrw==0.4
pep8==1.7.1
pickleshare==0.7.4
Pillow==4.0.0
pkginfo==1.4.2
pluggy==0.6.0
ply==3.11
pptx-template==0.2.6
probableparsing==0.0.1
probablepeople==0.5.4
prompt-toolkit==1.0.15
psutil==5.4.5
psycopg2==2.7.5
py==1.5.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.6.4
pycurl==7.43.0.1
PyDrive==1.3.1
pyexcel==0.5.8
pyexcel-io==0.5.7
pyexcel-xls==0.5.7
pyexcel-xlsx==0.5.6
pyflakes==1.6.0
Pygments==2.2.0
pygsheet==0.1.19b19
pygsheets==1.1.4
PyHamcrest==1.9.0
PyJWT==1.6.1
pylint==1.8.4
pyodbc==4.0.23
pyOpenSSL==17.5.0
pyparsing==2.2.0
PyPDF2==1.26.0
Pyphen==0.9.4
pyRFC3339==1.1
PySocks==1.6.8
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-crfsuite==0.9.5
python-dateutil==2.7.2
python-pptx==0.6.6
python3-openid==3.1.0
pytz==2018.4
PyWavelets==0.5.2
pywin32==223
pywinpty==0.5.1
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
rope==0.10.7
rsa==3.4.2
ruamel-yaml==0.15.35
s3transfer==0.1.13
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8.1
selenium==3.13.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
snowflake-connector-python==1.6.5
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sortedcollections==0.6.1
sortedcontainers==1.5.10
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.7
statsmodels==0.9.0
sympy==1.1.1
tables==3.4.3
tblib==1.3.2
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
texttable==1.2.1
tinycss2==0.6.1
toolz==0.9.0
tornado==5.0.2
tqdm==4.23.4
traitlets==4.3.2
Twisted==18.7.0
txaio==18.7.1
typing==3.6.4
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.2.0
virtualenv-clone==0.3.0
watson-developer-cloud==1.7.0
wcwidth==0.1.7
WeasyPrint==0.42.3
webcolors==1.8.1
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==3.3.1
widgetsnbextension==3.2.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xhtml2pdf==0.2.2
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.8
xlwt==1.3.0
xmltodict==0.11.0
zict==0.1.3
zope.component==4.4.1
zope.event==4.3.0
zope.interface==4.5.0
